### 2023-03-04 #5212 @editorialbot
[PRE REVIEW]: graphlayouts: Layout algorithms for network visualizations in R
**Submitting author:** @schochastics (<a href="http://orcid.org/0000-0003-2952-4812">David Schoch</a>)
**Repository:** https://github.com/schochastics/graphlayouts
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.4
