### 2023-03-04 #5212 @danielskatz
@schochastics - thanks for your submission.  I'm going to ask the JOSS editorial team to discuss if it is research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, as we often have questions about software that focuses on visualization, and I can't tell how much of this is visualization vs analysis, which is typically in-scope for JOSS.
