### 2023-03-09 #5238 @editorialbot
[REVIEW]: graphlayouts: Layout algorithms for network visualizations in R
**Submitting author:** @schochastics (<a href="http://orcid.org/0000-0003-2952-4812">David Schoch</a>)
**Repository:** https://github.com/schochastics/graphlayouts
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
