### 2023-03-14 schochastics/graphlayouts#51 @bpbond
ggplot2 warning when running README example
Running the first example in the README produces a warning from ggplot2:

```r
> ggraph(pa,layout = "nicely")+
+     geom_edge_link0(width=0.2,colour="grey")+
+     geom_node_point(col="black",size=0.3)+
+     theme_graph()
Warning message:
Using the `size` aesthetic in this geom was deprecated in ggplot2 3.4.0.
ℹ Please use `linewidth` in the `default_aes` field and elsewhere instead.
This warning is displayed once every 8 hours.
Call `lifecycle::last_lifecycle_warnings()` to see where this warning was
generated.
```

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
