### 2023-03-14 schochastics/graphlayouts#52 @bpbond
README example error: oaqc needed
The example that begins with
```r
bb <- layout_as_backbone(g,keep=0.4)
```
produces an error:
```
Error: oaqc is needed for this function to work. Please install it.
```
This isn't very clear. I assume that `oaqc` is a package, and available on CRAN? Is there a reason it's not listed in the package DESCRIPTION file under `Suggests:` or `Imports:`?

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
