### 2023-03-14 schochastics/graphlayouts#53 @bpbond
DESCRIPTION text tweak
The last sentence in the Description: field is grammatically awkward:

>    Some more specific algorithms allow to emphasize hidden group structures in networks or focus on specific nodes.

Consider changing to "allow the user to".

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
