### 2023-03-14 schochastics/graphlayouts#54 @bpbond
dynamic layouts example
The "dynamic layouts" example in the README begins
```r
#gList is a list of longitudinal networks.

xy <- layout_as_dynamic(gList,alpha = 0.2)
pList <- vector("list",length(gList))
```

This is unclear; where does `gList` come from? Is there a way to make this example reproducible?

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
