### 2023-03-14 schochastics/graphlayouts#55 @bpbond
1:length constructs
I'm looking at the code for [layout_as_dynamic](https://github.com/schochastics/graphlayouts/blob/main/R/layout_dynamic.R) and see many loops similar to this:

```
  for(i in 1:length(gList)){
```
This is risky in R, because if `gList` is an empty list, the loop will execute twice!
```
> for(i in 1:length(list())) print(i)
[1] 1
[1] 0
```

In general it's safer to use R's `seq_along`.

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
