### 2023-03-14 schochastics/graphlayouts#56 @bpbond
Add contributing.md file?
It might be useful to [add a contributing.md file](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file).

This would also help check the "Community guidelines" box in https://github.com/openjournals/joss-reviews/issues/5238.
