### 2023-03-14 schochastics/graphlayouts#57 @bpbond
Clarify last sentence in statement of need
I found the last sentence of the statement of need

confusing. What's the relationship between `ggraph` and `graphlayouts`? Something seems missing here.

Minor: Line 23, "packages"

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
