### 2023-03-14 schochastics/graphlayouts#58 @bpbond
Vignette?
Is there a reason that https://www.mr.schochastics.net/material/netVizR/ couldn't become a built-in vignette? This would provide a more natural resource for R users.

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
