### 2023-03-14 schochastics/graphlayouts#59 @bpbond
Minor grammar notes
Re [the paper text](https://github.com/openjournals/joss-papers/blob/joss.05238/joss.05238/10.21105.joss.05238.pdf):

- [ ] Line 24, "layouts"
- [ ] L. 27, see #58
- [ ] L. 38, "requires the computation of the full"
- [ ] L. 43, "benchmark"

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
