### 2023-03-14 #5238 @bpbond
Hi @schochastics

This is an interesting package and nicely written but quite limited paper. I have and am continuing to open specific issues in the target repo, but here are some broader thoughts and suggestions:
* The paper currently doesn't say anything about how to install the software, or minimum version requirements, or dependencies. These are all important to specify. (This does appear in the [online non-R documentation](https://www.mr.schochastics.net/material/netVizR/#required-libraries) I see.)
* The relationship between `ggraph`, `igraph`, and `graphlayouts` is not always clear...I suspect that beginners will appreciate having this spelled out explicitly.
* More generally, a short paragraph about the necessity and utility of network graphs (kinds and volumes of data, etc) would be helpful to introduce readers to their role in data visualization.
* The package help pages are generally solid, thanks.
* The existing package tests are...somewhat minimal; most seem to be 'does it work under a single case'. Consider adding some edge/corner case testing, e.g. for empty lists - cf. https://github.com/schochastics/graphlayouts/issues/55
