### 2023-03-14 schochastics/graphlayouts#51 @schochastics
Yes this is a long standing issue of `ggraph` because of changes of `ggplot2`. Feel free to reopen but I close this here, given that it is an independent issue
