+++ README.Rmd
@@ -128,6 +128,12 @@ ggraph(g,layout = "stress")+
 The backbone layout helps to uncover potential group structures based on edge 
 embeddedness and puts more emphasis on this structure in the layout.
 
+To use the function, you need to install the package `oaqc`
+```{r install_oaqc, eval=FALSE}
+install.packages("oaqc")
+```
+
+
 ```{r backbone,eval=FALSE}
 bb <- layout_as_backbone(g,keep=0.4)
 E(g)$col <- F
