+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: graphlayouts
 Title: Additional Layout Algorithms for Network Visualizations
-Version: 0.8.4
+Version: 0.8.4.9000
 Authors@R: person("David", "Schoch", email = "david@schochastics.net", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2952-4812"))
 Description: Several new layout algorithms to visualize networks are provided which are not part of 'igraph'. 
