### 2023-03-14 schochastics/graphlayouts#52 @schochastics
hmm it is part of the DESCRIPTION file: https://github.com/schochastics/graphlayouts/blob/298ee887f6ddbf52b96fc009a7a0e9ae47337c05/DESCRIPTION#L19

I added the installation to the README
