+++ DESCRIPTION
@@ -5,7 +5,7 @@ Authors@R: person("David", "Schoch", email = "david@schochastics.net", role = c(
            comment = c(ORCID = "0000-0003-2952-4812"))
 Description: Several new layout algorithms to visualize networks are provided which are not part of 'igraph'. 
     Most are based on the concept of stress majorization by Gansner et al. (2004) <doi:10.1007/978-3-540-31843-9_25>. 
-    Some more specific algorithms allow to emphasize hidden group structures in networks or focus on specific nodes.
+    Some more specific algorithms allow the user to emphasize hidden group structures in networks or focus on specific nodes.
 URL: http://graphlayouts.schochastics.net/, https://github.com/schochastics/graphlayouts
 BugReports: https://github.com/schochastics/graphlayouts/issues
 Depends: R (>= 3.2.0)
