### 2023-03-14 7667e30
replaced 1:length with seq_along (#55)
