### 2023-03-14 82b985f
added contributing guide (fixes #56)
