### 2023-03-14 schochastics/graphlayouts#58 @schochastics
True. It is far too long though in my opinion for a vignette. Will think how I can include a condensed version somehow
