### 2023-03-14 0a0827b
fixed typos (#59)
