### 2023-03-14 schochastics/graphlayouts#60 @schochastics
better tests
add better tests

Noted as part of https://github.com/openjournals/joss-reviews/issues/5238
