### 2023-03-14 4ba372f
added a vignette (#58)
