+++ vignettes/graphlayouts.Rmd
@@ -1,0 +1,216 @@
+---
+title: "graphlayouts"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{graphlayouts}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.align = "center",
+  fig.width = 7,
+  fig.height = 7
+)
+```
+
+This package implements some graph layout algorithms that are not available in `igraph` or other packages.
+This vignette only introduces the basic functionality of the package and the most important layout functions. 
+A more in depth tutorial on network visualization in R using
+`graphlayouts` which introduces its full functionality can be found [online](https://www.mr.schochastics.net/material/netVizR/).
+
+```{r load_libs,message=FALSE,warning=FALSE}
+library(igraph)   
+library(ggraph)   
+library(graphlayouts)
+```
+
+
+## Implemented layout algorithms
+
+The package implements the following algorithms:
+
+- Stress majorization ([Paper](https://graphviz.gitlab.io/_pages/Documentation/GKN04.pdf))
+- Quadrilateral backbone layout ([Paper](https://jgaa.info/accepted/2015/NocajOrtmannBrandes2015.19.2.pdf))
+- flexible radial layouts ([Paper](https://jgaa.info/accepted/2011/BrandesPich2011.15.1.pdf))
+- sparse stress ([Paper](https://arxiv.org/abs/1608.08909))
+- pivot MDS ([Paper](https://kops.uni-konstanz.de/bitstream/handle/123456789/5741/bp_empmdsld_06.pdf?sequence=1&isAllowed=y))
+- dynamic layout for longitudinal data ([Paper](https://kops.uni-konstanz.de/bitstream/handle/123456789/20924/Brandes_209246.pdf?sequence=2))
+- spectral layouts (adjacency/Laplacian)
+- a simple multilevel layout
+- a layout algorithm using UMAP
+- group based centrality and focus layouts which keeps groups of nodes close in the same range on the concentric circle
+
+## Stress layout for connected network  
+
+The "stress" layout is always a safe layout choice since it is deterministic and produces nice plots for almost any graph. It is recommended to use it as a default choice. 
+
+```{r example}
+set.seed(666)
+pa <- sample_pa(1000,1,1,directed = F)
+
+ggraph(pa,layout = "stress")+
+  geom_edge_link0(width = 0.2,colour = "grey")+
+  geom_node_point(color = "black",size = 0.3)+
+  theme_graph()
+```
+
+
+## Stress layout for unconnected network
+
+Stress majorization also works for networks with several components. It relies 
+on a bin packing algorithm to efficiently put the components in a rectangle, rather than a circle.
+
+```{r example_un}
+set.seed(666)
+g <- disjoint_union(
+  sample_pa(10,directed = F),
+  sample_pa(20,directed = F),
+  sample_pa(30,directed = F),
+  sample_pa(40,directed = F),
+  sample_pa(50,directed = F),
+  sample_pa(60,directed = F),
+  sample_pa(80,directed = F)
+)
+
+ggraph(g, layout = "stress",bbox = 40) +
+  geom_edge_link0() +
+  geom_node_point() +
+  theme_graph()
+```
+
+## Backbone layout
+
+`layout_as_backbone()` is a layout algorithm that can help emphasize hidden group structures. To illustrate the performance of the algorithm, we create an artificial network with a subtle group structure using `sample_islands()` from `igraph`.
+
[…]
