+++ DESCRIPTION
@@ -28,3 +28,4 @@ LinkingTo:
     RcppArmadillo
 RoxygenNote: 7.2.1
 Roxygen: list(markdown = TRUE)
+VignetteBuilder: knitr
