### 2023-03-14 schochastics/graphlayouts#58 @schochastics
I added a rudimentary vignette which introduces the main layout algorithm and points to the online tutorial. The tutorial depends on too many packages which I would need to add to the package. Hence the very boiled down version
