### 2023-03-14 #5238 @schochastics
@bpbond

- Installation details were added. I did not include version numbers because there is no minimum requirement in this specific case. In the tutorial, it is only present because of the interplay of `ggraph` and `ggforce`.

- I tried to be more explicit in the paper what igraph and ggraph are and how they relate with graphlayouts. In my opinion, the plotting section serves the purpose of explaining the relation, i.e. that both can use the layout algorithms for plotting. Happy to add more explanations if you think it is needed.

-   I wasnt exactly sure how to interpret your comment of adding a "short paragraph about the necessity and utility of network graphs"

- Thanks for calling out the tests. I added some more that also test edge/corner cases. Happy to add even more if you think it is necessary
