### 2023-03-15 c5dedb1
added test (#60)
