### 2023-03-15 bda950a
fixed igraph layout function tests
