+++ tests/testthat/test-layouts.R
@@ -1,13 +1,12 @@
-test_that("igraph functions should never be called", {
+test_that("igraph function calls work", {
   g <- igraph::make_full_graph(10,FALSE)
-  expect_error(layout_igraph_centrality(g))
-  expect_error(layout_igraph_centrality_group(g))
-  expect_error(layout_igraph_constrained_stress(g))
-  expect_error(layout_igraph_eigen(g))
-  expect_error(layout_igraph_focus(g))
-  expect_error(layout_igraph_focus_group(g))
-  expect_error(layout_igraph_multilevel(g))
-  expect_error(layout_igraph_pmds(g))
-  expect_error(layout_igraph_sparse_stress(g))
-  expect_error(layout_igraph_stress(g))
+  igraph::V(g)$lvl <- rep(c(1,2),each = 5)
+  expect_no_error(layout_igraph_centrality(g,cent = igraph::degree(g)))
+  expect_no_error(layout_igraph_constrained_stress(g,fixdim="x",coord=rep(1,10)))
+  expect_no_error(layout_igraph_eigen(g))
+  expect_no_error(layout_igraph_focus(g,v = 1))
+  expect_no_error(layout_igraph_multilevel(g))
+  expect_no_error(layout_igraph_pmds(g,pivots = 5))
+  expect_no_error(layout_igraph_sparse_stress(g,pivots = 5))
+  expect_no_error(layout_igraph_stress(g))
 })
