+++ DESCRIPTION
@@ -26,6 +26,6 @@ Suggests:
 LinkingTo: 
     Rcpp,
     RcppArmadillo
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Roxygen: list(markdown = TRUE)
 VignetteBuilder: knitr
