### 2023-03-15 schochastics/graphlayouts#58 @bpbond
```r
> devtools::install_github("schochastics/graphlayouts")
> library(graphlayouts)
> vignette(package="graphlayouts")
no vignettes found
```

Hmm. I looked through the [vignette Rmarkdown file](https://github.com/schochastics/graphlayouts/blob/main/vignettes/graphlayouts.Rmd) and it looks great but not sure why it's not visible when I install as above. Anyway, closing as addressed—thanks!—but please double-check it's available to users.
