### 2023-03-15 #5238 @bpbond
@schochastics Thanks for your changes to the paper, vignette addition, new tests (great!), and other clarifications. Nice job!

@crvernon These changes fully address my feedback above and I have no further suggestions.
