### 2023-03-22 #5238 @crvernon
:mega:  Mid-week Rally! :mega:

Looks like you guys are making great progress!  I see that @bpbond and @schochastics have wrapped up that portion of the review!  Nice work!

@ndjackso thanks for setting up your checklist as well!  Please let me know if you have any questions as you progress through the process!
