### 2023-04-26 #5238 @ndjackso
Hi, @schochastics.

Thank you for leading the development of the graphlayouts package to support network visualization. Significant package documentation and working examples have been made available through Githab, CRAN, and a [website](https://www.mr.schochastics.net/material/netVizR/). The package has had multiple releases and has been cited by others. As part of the review, I also tried using the cowplot package as an alternative to combining visualizations generated using graphlayouts (no issues) and other color palettes from the RColorBrewer and viridis packages as these are common packages for my data visualization workflow. The submitted paper provides succinct information related to the package. However, there were a few issues I encountered when replicating examples and reviewing the manuscript as well as package documentation. Some suggestions for your consideration:

- Beyond the “Summary”, the other sections in the software paper could use more detail.
   - The “Statement of Need” section makes reference to igragh and ggraph but there could be an expansion of content to situate graphlayouts relative to current static (e.g., `network`, `sna`, `statnet`, etc.) and dynamic (e.g., `networkD3` and `VizNetwork`) network visualization packages in R.
   - The “Use for plotting” section may be redundant as the plotting capabilities are demonstrated in the subsequent sections and code. Consider removing this section or add additional content to distinguish it further from the other sections
   - The “Overview of algorithms” section shows some of the visualization functions but the descriptions are minimal. There is also unequal coverage of the algorithms presented in the Github repo.  Consider adding more detail for all algorithms.
- There is a license available but the type is not specified. Consider adding this information to the LICENSE file
- Some packages are checked during the installation but others are not. When I installed the package using the dev version, I was prompted to update `rlang`, `RccpArmadillo`, and the `igraph` packages. I did not have `oaqc` installed, and I was not flagged for this until working through the provided examples. Consider adding all packages as part of the check when installing the dev version.
- I received a warning when re-creating the first ggraph example from the “Use for plotting” section:

>  “Warning: Using the `size` aesthetic in this geom was deprecated in ggplot2 3.4.0. Please use `linewidth` in the `default_aes` field and elsewhere instead.”

Consider reviewing functions to ensure deprecated features from dependent packages are not being used in the current package version

- I received a warning when changing color palettes in the Backbone layout example:

> “Warning: n too large, allowed maximum for palette Dark2 is 8 Returning the palette you asked for with that many colors”.

 I also encountered this warning with Set2, Pastel2, and Accent palettes in RColorBrewer. This issue appears to stem from the palette choice and not the graphlayouts package. One of the package’s strengths is being able to visualize very large networks. A couple of thoughts to overcome the warning:

   - Consider adding a recommended of a color palette that works for large networks as part of the warning
   - Considering including a default color palette that works for large n to override limited palettes

I'll create separate issues for this considerations.
