### 2023-04-26 schochastics/graphlayouts#61 @ndjackso
Paper/Statement of need: other network packages
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
