### 2023-04-26 schochastics/graphlayouts#62 @ndjackso
Paper/Use for plotting: section necessary
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
