### 2023-04-26 schochastics/graphlayouts#63 @ndjackso
Paper/Overview of algorithms: more details
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
