### 2023-04-26 schochastics/graphlayouts#64 @ndjackso
License: specify which type
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
