### 2023-04-26 schochastics/graphlayouts#65 @ndjackso
Installation: check all dependencies
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
