### 2023-04-26 schochastics/graphlayouts#66 @ndjackso
Using Deprecated gglot2 features
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
