### 2023-04-26 schochastics/graphlayouts#67 @ndjackso
Color palette limitations
Issue related to [openjournals/joss-reviews#5238](https://github.com/openjournals/joss-reviews/issues/5238)
