### 2023-04-26 schochastics/graphlayouts#67 @schochastics
>   I received a warning when changing color palettes in the Backbone layout example:
    “Warning: n too large, allowed maximum for palette Dark2 is 8 Returning the palette you asked for with that many colors”.
I also encountered this warning with Set2, Pastel2, and Accent palettes in RColorBrewer. This issue appears to stem from the palette choice and not the graphlayouts package. One of the package’s strengths is being able to visualize very large networks. A couple of thoughts to overcome the warning:
   Consider adding a recommended of a color palette that works for large networks as part of the warning
    Considering including a default color palette that works for large n to override limited palettes

Yeah Set1 is afaik the only discrete color palette from RColorBrewer that allows 9 colors. I am reluctant to change anything here though, because I do not think this is an issue that `graphlayouts` needs to solve. Color palettes are to some extend subjective, and I do not want to define a default here. let me know if you disagree strongly here and I can implement something. Otherwise just close the issue
