+++ paper.bib
@@ -67,3 +67,22 @@
   organization={Springer},
   doi={10.1007/978-3-319-50106-2_2}
 }
+
+@article{handcock2008statnet,
+  title={statnet: Software tools for the representation, visualization, analysis and simulation of network data},
+  author={Handcock, Mark S and Hunter, David R and Butts, Carter T and Goodreau, Steven M and Morris, Martina},
+  journal={Journal of statistical software},
+  volume={24},
+  number={1},
+  pages={1548},
+  year={2008},
+  publisher={NIH Public Access}
+}
+
+ @manual{visNetwork2022,
+    title = {visNetwork: Network Visualization using 'vis.js' Library},
+    author = {Almende, BV and Thieurmel, Benoit and Robert, Titouan},
+    year = {2022},
+    note = {R package version 2.1.2},
+    url = {https://CRAN.R-project.org/package=visNetwork},
+  }
