### 2023-04-26 schochastics/graphlayouts#61 @schochastics
>The “Statement of Need” section makes reference to igragh and ggraph but there could be an expansion of content to situate graphlayouts relative to current static (e.g., network, sna, statnet, etc.) and dynamic (e.g., networkD3 and VizNetwork) network visualization packages in R.

Thanks for the suggestion! I included those packages in the statement of need
