### 2023-04-26 schochastics/graphlayouts#62 @schochastics
> The “Use for plotting” section may be redundant as the plotting capabilities are demonstrated in the subsequent sections and code. Consider removing this section or add additional content to distinguish it further from the other sections

agreed and was deleted
