### 2023-04-26 schochastics/graphlayouts#63 @schochastics
> The “Overview of algorithms” section shows some of the visualization functions but the descriptions are minimal. There is also unequal coverage of the algorithms presented in the Github repo. Consider adding more detail for all algorithms.

This is indeed a good point. My decision to keep it minimal are guided by the journals guidelines. The paper is not supposed to have a tutorial character. This is why I opted for only shortly presenting the important algorithms. I clarified this a bit more now at the beginning of the section and explicitly point to more detailed resources (also on github)
