### 2023-04-26 schochastics/graphlayouts#67 @ndjackso
I was on the fence about including the warning as an issue recognizing that it might be beyond scope of the package. Color palettes are subjective. However, the package integrates well with other visualization packages where users can find something that works for them. Glad to have had this discussion, and I'm okay with closing the issue.
