### 2023-04-26 #5238 @ndjackso
Hi, @schochastics.

Thanks for responding to the issues! The software paper additions look good, and I appreciate your thoughtful responses to the more open-ended issues.
