+++ cran-comments.md
@@ -8,7 +8,7 @@ no user facing changes
 All dependencies passed R CMD check
 
 ## Test environments
-* ubuntu 20.04, R 4.2.2
+* ubuntu 22.04, R 4.2.2
 * win-builder (devel and release)
 
 ## R CMD check results
