+++ README.md
@@ -11,6 +11,7 @@ status](https://www.r-pkg.org/badges/version/graphlayouts)](https://cran.r-proje
 Downloads](https://cranlogs.r-pkg.org/badges/grand-total/graphlayouts)](https://CRAN.R-project.org/package=graphlayouts)
 [![Codecov test
 coverage](https://codecov.io/gh/schochastics/graphlayouts/branch/main/graph/badge.svg)](https://app.codecov.io/gh/schochastics/graphlayouts?branch=main)
+[![Zenodo](https://zenodo.org/badge/DOI/10.5281/zenodo.7870213.svg)](https://doi.org/10.5281/zenodo.7870213)
 
 This package implements some graph layout algorithms that are not
 available in `igraph`.
