### 2023-04-27 #5238 @crvernon
:wave: @schochastics - we are almost there! Your paper and software are looking great!  Please address the following item:

- [ ] Add the DOI to the Handcock et al. 2008 citation.  I believe this is it `10.18637/jss.v024.i01` but please double check.

Once this change has been made we will move to the final steps of setting up the archive for your new release.

So here is what we have left to do:

- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ] Please list the DOI of the archived version here
