+++ paper.bib
@@ -76,7 +76,8 @@
   number={1},
   pages={1548},
   year={2008},
-  publisher={NIH Public Access}
+  publisher={NIH Public Access},
+  doi = {10.18637/jss.v024.i01}
 }
 
  @manual{visNetwork2022,
