### 2023-04-27 #5238 @schochastics
@crvernon

I added the missing doi and here are the links:

- tagged release: v1.0.0 [link](https://github.com/schochastics/graphlayouts/releases/tag/v1.0.0)
- Zenodo archive: https://zenodo.org/record/7870213
- DOI: https://doi.org/10.5281/zenodo.7870213
