### 2023-04-29 #5238 @arfon
@schochastics – I found a tiny issue with the formatting of your paper. Could you please merge this PR: https://github.com/schochastics/graphlayouts/pull/68 ?
